chore(backend): change name of query

+++ apps/backend/src/models.rs
@@ -175,7 +175,7 @@ pub mod media {
     pub struct GenreListItem {
         pub id: i32,
         pub name: String,
-        pub num_items: Option<i32>,
+        pub num_items: Option<i64>,
     }
 
     #[derive(Debug, Serialize, Deserialize, SimpleObject, Clone, FromQueryResult)]
